api/requirements.txt
flask
flask-restful
flask-script
flask-sqlalchemy
flask-migrate
flask-marshmallow
psycopg2-binary
marshmallow-sqlalchemy
marshmallow
python-dotenv
numpy
pandas
newsapi-python
ibm-watson
#google-cloud-translate
git+https://github.com/qmeeus/textrank.git
gunicorn
logzero
feedparser